backend/pyproject.toml
[build-system]
requires = ["setuptools>=69.0"]
build-backend = "setuptools.build_meta"

[project]
name = "smart-pot-backend"
version = "0.1.0"
description = "FastAPI service that exposes Smart Pot HTTP APIs."
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "Smart Pot Team" }]
dependencies = [
    "fastapi>=0.110.0,<1.0",
    "uvicorn[standard]>=0.25.0,<0.30",
    "psycopg2-binary>=2.9.9,<3.0",
    "passlib[argon2]>=1.7.4,<2.0",
    "argon2-cffi>=23.1.0,<24.0",
    "PyJWT>=2.8.0,<3.0",
    "paho-mqtt>=1.6.1,<2.0",
    "pydantic[email]>=2.6.1,<3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.4,<9.0",
    "httpx>=0.26.0,<0.28",
    "ruff>=0.2.1,<0.7",
]

[tool.ruff]
line-length = 100
target-version = "py311"
